\name{dmy}
\alias{dmy}
\title{Wrapper function around lubridate::dmy that gives feedback about parsed variable}
\usage{
dmy(.data, ...)
}
\arguments{
\item{.data}{see \link[lubridate:dmy]{dmy}}

\item{...}{see \link[lubridate:dmy]{dmy}}
}
\value{
see \link[lubridate:dmy]{dmy}
}
\description{
Wrapper function around lubridate::dmy that gives feedback about parsed variable
}
