\name{dmy_h}
\alias{dmy_h}
\title{Wrapper function around lubridate::dmy_h that gives feedback about parsed variable}
\usage{
dmy_h(.data, ...)
}
\arguments{
\item{.data}{see \link[lubridate:dmy_h]{dmy_h}}

\item{...}{see \link[lubridate:dmy_h]{dmy_h}}
}
\value{
see \link[lubridate:dmy_h]{dmy_h}
}
\description{
Wrapper function around lubridate::dmy_h that gives feedback about parsed variable
}
