\name{dmy_hms}
\alias{dmy_hms}
\title{Wrapper function around lubridate::dmy_hms that gives feedback about parsed variable}
\usage{
dmy_hms(.data, ...)
}
\arguments{
\item{.data}{see \link[lubridate:dmy_hms]{dmy_hms}}

\item{...}{see \link[lubridate:dmy_hms]{dmy_hms}}
}
\value{
see \link[lubridate:dmy_hms]{dmy_hms}
}
\description{
Wrapper function around lubridate::dmy_hms that gives feedback about parsed variable
}
