\name{hm}
\alias{hm}
\title{Wrapper function around lubridate::hm that gives feedback about parsed variable}
\usage{
hm(.data, ...)
}
\arguments{
\item{.data}{see \link[lubridate:hm]{hm}}

\item{...}{see \link[lubridate:hm]{hm}}
}
\value{
see \link[lubridate:hm]{hm}
}
\description{
Wrapper function around lubridate::hm that gives feedback about parsed variable
}
