\name{hms}
\alias{hms}
\title{Wrapper function around lubridate::hms that gives feedback about parsed variable}
\usage{
hms(.data, ...)
}
\arguments{
\item{.data}{see \link[lubridate:hms]{hms}}

\item{...}{see \link[lubridate:hms]{hms}}
}
\value{
see \link[lubridate:hms]{hms}
}
\description{
Wrapper function around lubridate::hms that gives feedback about parsed variable
}
