\name{mdy_h}
\alias{mdy_h}
\title{Wrapper function around lubridate::mdy_h that gives feedback about parsed variable}
\usage{
mdy_h(.data, ...)
}
\arguments{
\item{.data}{see \link[lubridate:mdy_h]{mdy_h}}

\item{...}{see \link[lubridate:mdy_h]{mdy_h}}
}
\value{
see \link[lubridate:mdy_h]{mdy_h}
}
\description{
Wrapper function around lubridate::mdy_h that gives feedback about parsed variable
}
