\name{mdy_hms}
\alias{mdy_hms}
\title{Wrapper function around lubridate::mdy_hms that gives feedback about parsed variable}
\usage{
mdy_hms(.data, ...)
}
\arguments{
\item{.data}{see \link[lubridate:mdy_hms]{mdy_hms}}

\item{...}{see \link[lubridate:mdy_hms]{mdy_hms}}
}
\value{
see \link[lubridate:mdy_hms]{mdy_hms}
}
\description{
Wrapper function around lubridate::mdy_hms that gives feedback about parsed variable
}
