\name{ms}
\alias{ms}
\title{Wrapper function around lubridate::ms that gives feedback about parsed variable}
\usage{
ms(.data, ...)
}
\arguments{
\item{.data}{see \link[lubridate:ms]{ms}}

\item{...}{see \link[lubridate:ms]{ms}}
}
\value{
see \link[lubridate:ms]{ms}
}
\description{
Wrapper function around lubridate::ms that gives feedback about parsed variable
}
