\name{my}
\alias{my}
\title{Wrapper function around lubridate::my that gives feedback about parsed variable}
\usage{
my(.data, ...)
}
\arguments{
\item{.data}{see \link[lubridate:my]{my}}

\item{...}{see \link[lubridate:my]{my}}
}
\value{
see \link[lubridate:my]{my}
}
\description{
Wrapper function around lubridate::my that gives feedback about parsed variable
}
