\name{myd}
\alias{myd}
\title{Wrapper function around lubridate::myd that gives feedback about parsed variable}
\usage{
myd(.data, ...)
}
\arguments{
\item{.data}{see \link[lubridate:ymd]{myd}}

\item{...}{see \link[lubridate:ymd]{myd}}
}
\value{
see \link[lubridate:ymd]{myd}
}
\description{
Wrapper function around lubridate::myd that gives feedback about parsed variable
}
