\name{ydm_h}
\alias{ydm_h}
\title{Wrapper function around lubridate::ydm_h that gives feedback about parsed variable}
\usage{
ydm_h(.data, ...)
}
\arguments{
\item{.data}{see \link[lubridate:ydm_h]{ydm_h}}

\item{...}{see \link[lubridate:ydm_h]{ydm_h}}
}
\value{
see \link[lubridate:ydm_h]{ydm_h}
}
\description{
Wrapper function around lubridate::ydm_h that gives feedback about parsed variable
}
