\name{ydm_hms}
\alias{ydm_hms}
\title{Wrapper function around lubridate::ydm_hms that gives feedback about parsed variable}
\usage{
ydm_hms(.data, ...)
}
\arguments{
\item{.data}{see \link[lubridate:ydm_hms]{ydm_hms}}

\item{...}{see \link[lubridate:ydm_hms]{ydm_hms}}
}
\value{
see \link[lubridate:ydm_hms]{ydm_hms}
}
\description{
Wrapper function around lubridate::ydm_hms that gives feedback about parsed variable
}
