\name{year}
\alias{year}
\title{Wrapper function around lubridate::year that gives feedback about parsed variable}
\usage{
year(.data, ...)
}
\arguments{
\item{.data}{see \link[lubridate:year]{year}}

\item{...}{see \link[lubridate:year]{year}}
}
\value{
see \link[lubridate:year]{year}
}
\description{
Wrapper function around lubridate::year that gives feedback about parsed variable
}
