\name{ym}
\alias{ym}
\title{Wrapper function around lubridate::ym that gives feedback about parsed variable}
\usage{
ym(.data, ...)
}
\arguments{
\item{.data}{see \link[lubridate:ym]{ym}}

\item{...}{see \link[lubridate:ym]{ym}}
}
\value{
see \link[lubridate:ym]{ym}
}
\description{
Wrapper function around lubridate::ym that gives feedback about parsed variable
}
