\name{ymd}
\alias{ymd}
\title{Wrapper function around lubridate::ymd that gives feedback about parsed variable}
\usage{
ymd(.data, ...)
}
\arguments{
\item{.data}{see \link[lubridate:ymd]{ymd}}

\item{...}{see \link[lubridate:ymd]{ymd}}
}
\value{
see \link[lubridate:ymd]{ymd}
}
\description{
Wrapper function around lubridate::ymd that gives feedback about parsed variable
}
