\name{ymd_h}
\alias{ymd_h}
\title{Wrapper function around lubridate::ymd_h that gives feedback about parsed variable}
\usage{
ymd_h(.data, ...)
}
\arguments{
\item{.data}{see \link[lubridate:ymd_h]{ymd_h}}

\item{...}{see \link[lubridate:ymd_h]{ymd_h}}
}
\value{
see \link[lubridate:ymd_h]{ymd_h}
}
\description{
Wrapper function around lubridate::ymd_h that gives feedback about parsed variable
}
