\name{ymd_hms}
\alias{ymd_hms}
\title{Wrapper function around lubridate::ymd_hms that gives feedback about parsed variable}
\usage{
ymd_hms(.data, ...)
}
\arguments{
\item{.data}{see \link[lubridate:ymd_hms]{ymd_hms}}

\item{...}{see \link[lubridate:ymd_hms]{ymd_hms}}
}
\value{
see \link[lubridate:ymd_hms]{ymd_hms}
}
\description{
Wrapper function around lubridate::ymd_hms that gives feedback about parsed variable
}
