\name{yq}
\alias{yq}
\title{Wrapper function around lubridate::yq that gives feedback about parsed variable}
\usage{
yq(.data, ...)
}
\arguments{
\item{.data}{see \link[lubridate:yq]{yq}}

\item{...}{see \link[lubridate:yq]{yq}}
}
\value{
see \link[lubridate:yq]{yq}
}
\description{
Wrapper function around lubridate::yq that gives feedback about parsed variable
}
