\name{mdy}
\alias{mdy}
\title{Wrapper function around lubridate::mdy that gives feedback about parsed variable}
\usage{
mdy(.data, ...)
}
\arguments{
\item{.data}{see \link[lubridate:mdy]{mdy}}

\item{...}{see \link[lubridate:mdy]{mdy}}
}
\value{
see \link[lubridate:mdy]{mdy}
}
\description{
Wrapper function around lubridate::mdy that gives feedback about parsed variable
}
