\name{ydm}
\alias{ydm}
\title{Wrapper function around lubridate::ydm that gives feedback about parsed variable}
\usage{
ydm(.data, ...)
}
\arguments{
\item{.data}{see \link[lubridate:ymd]{ydm}}

\item{...}{see \link[lubridate:ymd]{ydm}}
}
\value{
see \link[lubridate:ymd]{ydm}
}
\description{
Wrapper function around lubridate::ydm that gives feedback about parsed variable
}
